requirements.txt
blinker==1.4
click==8.0.4
Flask==2.0.3
Flask-Mail==0.9.1
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
python-dateutil==2.8.2
python-dotenv==0.19.2
six==1.16.0
uWSGI==2.0.20
Werkzeug==2.0.3
